+++ jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-core/src/main/java/jdplus/toolkit/base/core/math/linearfilters/CrossValidation.java
@@ -18,7 +18,6 @@ package jdplus.toolkit.base.core.math.linearfilters;
 
 import jdplus.toolkit.base.api.data.DoubleSeq;
 import java.util.function.IntFunction;
-import jdplus.toolkit.base.core.math.linearfilters.SymmetricFilter;
 
 /**
  *