+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- ![STAT] ![UI] Add cross-validation algorithm for seasonal filter selection in the X-11 algorithm, including updates to core routines and the user interface
+- ![STAT] Add frequency in SA specifications
+
 ### Changed
 
 - ![IO] Improve I/O of CSV files [#163](https://github.com/jdemetra/jd3-main/issues/163)